requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
dask==2023.4.0
filelock==3.12.0
fonttools==4.39.3
fsspec==2023.4.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
Jinja2==3.1.2
kiwisolver==1.4.4
locket==1.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
nd2==0.5.3
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
partd==1.4.0
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.29.0
resource-backed-dask-array==0.1.0
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@567662b0fd33ca4b022d94d3b8de896628cd32dd
six==1.16.0
sympy==1.11.1
toolz==0.12.0
torch==2.0.0
torchvision==0.15.1
typing_extensions==4.5.0
urllib3==1.26.15
zipp==3.15.0
